ok so i fixed the double cout issue

+++ app/include/file.hpp
@@ -4,7 +4,7 @@
 
 namespace File {
 	std::string getFile(std::string path);
-	int sizeOfFile(std::string path);
+	unsigned long long sizeOfFile(std::string path);
 	std::vector<std::string> getThingsInFolder(std::string path);
 	std::string getMIMEType(std::string path);
 	std::string getChunkFromFile(std::string path, int chunk);
